ref(samplers): generic spawning
