+++ src/samplers/disk/mod.rs
@@ -66,22 +66,6 @@ impl Sampler for Disk {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().disk().enabled() {
-            if let Ok(mut sampler) = Self::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = sampler.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize disk sampler");
-            } else {
-                error!("failed to initialize disk sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
