+++ src/samplers/http/mod.rs
@@ -48,22 +48,6 @@ impl Sampler for Http {
         Ok(ret)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().http().enabled() {
-            if let Ok(mut sampler) = Self::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = sampler.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize http sampler");
-            } else {
-                error!("failed to initialize http sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
