+++ src/samplers/interrupt/mod.rs
@@ -62,22 +62,6 @@ impl Sampler for Interrupt {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().interrupt().enabled() {
-            if let Ok(mut interrupt) = Interrupt::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = interrupt.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize interrupt sampler");
-            } else {
-                error!("failed to initialize interrupt sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
