+++ src/samplers/krb5kdc/mod.rs
@@ -95,27 +95,6 @@ impl Sampler for Krb5kdc {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().krb5kdc().enabled() {
-            match Self::new(common.clone()) {
-                Ok(mut sampler) => {
-                    common.runtime().spawn(async move {
-                        loop {
-                            let _ = sampler.sample().await;
-                        }
-                    });
-                }
-                Err(e) => {
-                    if !common.config.fault_tolerant() {
-                        fatal!("failed to initialize krb5kdc sampler {}", e);
-                    } else {
-                        error!("failed to initialize krb5kdc sampler {}", e);
-                    }
-                }
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
