+++ src/samplers/memcache/mod.rs
@@ -56,22 +56,6 @@ impl Sampler for Memcache {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().memcache().enabled() {
-            if let Ok(mut sampler) = Self::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = sampler.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize memcache sampler");
-            } else {
-                error!("failed to initialize memcache sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
