+++ src/samplers/memory/mod.rs
@@ -48,22 +48,6 @@ impl Sampler for Memory {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().memory().enabled() {
-            if let Ok(mut sampler) = Self::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = sampler.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize memory sampler");
-            } else {
-                error!("failed to initialize memory sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
