+++ src/samplers/network/mod.rs
@@ -63,22 +63,6 @@ impl Sampler for Network {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().network().enabled() {
-            if let Ok(mut sampler) = Self::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = sampler.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize network sampler");
-            } else {
-                error!("failed to initialize network sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
