+++ src/samplers/process/mod.rs
@@ -57,22 +57,6 @@ impl Sampler for Process {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().process().enabled() {
-            if let Ok(mut sampler) = Self::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = sampler.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize process sampler");
-            } else {
-                error!("failed to initialize process sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
