+++ src/samplers/scheduler/mod.rs
@@ -88,22 +88,6 @@ impl Sampler for Scheduler {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().scheduler().enabled() {
-            if let Ok(mut sampler) = Self::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = sampler.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize scheduler sampler");
-            } else {
-                error!("failed to initialize scheduler sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
