+++ src/samplers/tcp/mod.rs
@@ -62,22 +62,6 @@ impl Sampler for Tcp {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().tcp().enabled() {
-            if let Ok(mut sampler) = Self::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = sampler.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize tcp sampler");
-            } else {
-                error!("failed to initialize tcp sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
