+++ src/samplers/xfs/mod.rs
@@ -57,22 +57,6 @@ impl Sampler for Xfs {
         Ok(sampler)
     }
 
-    fn spawn(common: Common) {
-        if common.config().samplers().xfs().enabled() {
-            if let Ok(mut sampler) = Self::new(common.clone()) {
-                common.runtime().spawn(async move {
-                    loop {
-                        let _ = sampler.sample().await;
-                    }
-                });
-            } else if !common.config.fault_tolerant() {
-                fatal!("failed to initialize xfs sampler");
-            } else {
-                error!("failed to initialize xfs sampler");
-            }
-        }
-    }
-
     fn common(&self) -> &Common {
         &self.common
     }
